requirements.txt
branca==0.8.0
certifi==2024.8.30
charset-normalizer==3.3.2
folium==0.17.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==2.1.1
python-dotenv==1.0.1
requests==2.32.3
scapy==2.6.0
urllib3==2.2.3
xyzservices==2024.9.0
